Strictly make the output conform to JUnit.xsd
While online sources seem to disagree on the exact details of the format, this
schema is what is referenced by the Azure docs, and it seems like a solid
baseline. If it turns out that not all services/tools can consume this version
of the format we might have to implement switches for output variants.