Use Kaocha's hierarchy helpers for finding leaf types test
Leaf tests are tests that don't have children themselves, e.g. test vars,
or cucumber scenarios. They are at the bottom of the graph, as opposed to tests
that act as containers, like namespaces, cucumber features, or test suites.

We use keyword derivation to label test types as being of a leaf type. This can
be done both through Kaocha's hierarchy (the new way) or through the global
clojure hierarchy (the old way). kaocha-junit-xml was only checking for the old
way, causing issues with kaocha-cucumber.